# docs/swagger.yaml
definitions:
...
  domain.CreateEvent:
    properties:
      name:
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.Event:
    properties:
      createdAt:
        type: string
      ended:
        type: boolean
      id:
        type: integer
      invites:
        items:
          $ref: '#/definitions/domain.Invite'
        type: array
      members:
        items:
          $ref: '#/definitions/domain.EventMember'
        type: array
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/domain.Product'
        type: array
      updatedAt:
        type: string
    type: object
  domain.EventMember:
    properties:
      event:
        $ref: '#/definitions/domain.Event'
      event_id:
        type: integer
      id:
        type: integer
      role:
        $ref: '#/definitions/domain.Role'
      role_id:
        type: integer
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
    type: object
  domain.EventMembers:
    properties:
      role_id:
        type: integer
      role_name:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  domain.EventWithAdmin:
    properties:
      event:
        $ref: '#/definitions/domain.EventWithMembersCount'
      is_admin:
        type: boolean
    type: object
  domain.EventWithMembersCount:
    properties:
      count:
        type: integer
      event:
        $ref: '#/definitions/domain.Event'
    type: object
  domain.Invite:
    properties:
      event:
        $ref: '#/definitions/domain.Event'
      event_id:
        type: integer
      expires_at:
        type: string
      id:
        type: string
    type: object
  domain.InviteLink:
    properties:
      expires_at:
        type: string
      link:
        type: string
    type: object
  domain.Product:
    properties:
      amount:
        type: integer
      bought:
        type: boolean
      consumer:
        $ref: '#/definitions/domain.User'
      consumer_id:
        type: integer
      createdAt:
        type: string
      equal:
        type: boolean
      event:
        $ref: '#/definitions/domain.Event'
      event_id:
        type: integer
      id:
        type: integer
      members:
        items:
          $ref: '#/definitions/domain.ProductMember'
        type: array
      name:
        type: string
      price:
        type: number
      updatedAt:
        type: string
    type: object
  domain.ProductMember:
    properties:
      amount:
        type: number
      id:
        type: integer
      product:
        $ref: '#/definitions/domain.Product'
      product_id:
        type: integer
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
    type: object
  domain.Role:
    properties:
      id:
        type: integer
      members:
        items:
          $ref: '#/definitions/domain.EventMember'
        type: array
      name:
        type: string
    type: object
  domain.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  domain.Token:
    properties:
      createdAt:
        type: string
      expires_in:
        type: string
      refresh:
        type: string
      revoke:
        type: boolean
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
    type: object
  domain.UpdateEvent:
    properties:
      ended:
        type: boolean
      id:
        type: integer
      name:
        type: string
    type: object
...
  domain.User:
    properties:
      bank:
        type: string
      createdAt:
        type: string
      events:
        items:
          $ref: '#/definitions/domain.EventMember'
        type: array
      id:
        type: integer
      phone_number:
        type: string
      tokens:
        items:
          $ref: '#/definitions/domain.Token'
        type: array
      updatedAt:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: API для tg Mini App "Приколы у Алёны"
  title: Rofl Bot API
  version: "1.0"
paths:
...
  /alena-rofl/event:
    post:
      consumes:
      - application/json
      description: Создает событие, при успехе возвращает id. авторизованный пользователь
        становится админом события.
      parameters:
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/domain.CreateEvent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание события
      tags:
      - event
    put:
      consumes:
      - application/json
      description: Можно изменить название или завершенность. оба параметра необязательны.
        !Изменить событие может только его админ!
      parameters:
      - description: Event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateEvent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление события
      tags:
      - event
  /alena-rofl/event/{id}:
    delete:
      consumes:
      - application/json
      description: по id удаляет событие. !Удалить событие может только его админ!
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление события
      tags:
      - event
    get:
      consumes:
      - application/json
      description: по id выдает событие. !не работает, если запрашивает не участник
        ивента!
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.EventWithAdmin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение события
      tags:
      - event
  /alena-rofl/event/{id}/admin:
    get:
      consumes:
      - application/json
      description: по id события выдает админа. !не работает, если запрашивает не
        участник ивента!
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение админа события
      tags:
      - event
  /alena-rofl/event/{id}/invite:
    post:
      consumes:
      - application/json
      description: Создает ссылку для приглашения в ивент с указанным id. не работает,
        если пользователь не участник ивента. Ссылка живет 24 часа, при повторном
        использовании старая ссылка заменяется на новую.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.InviteLink'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание ссылки для приглашения
      tags:
      - event
  /alena-rofl/event/{id}/leave:
    put:
      consumes:
      - application/json
      description: по id ивента удаляет из него авторизованного пользователя. !не
        работает, если пользователь не участник ивента!
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Покинуть ивент
      tags:
      - event
  /alena-rofl/event/{id}/members:
    get:
      consumes:
      - application/json
      description: по id ивента получает всех участников !не работает, если пользователь
        не участник ивента!
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.EventMembers'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение участников ивента
      tags:
      - event
  /alena-rofl/events:
    get:
      consumes:
      - application/json
      description: выдает все ивенты авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.EventWithMembersCount'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение всех ивентов текущего юзера
      tags:
      - event
  /alena-rofl/invite/{id}:
    post:
      consumes:
      - application/json
      description: по id invite-ссылки присоединяет авторизованного пользователя к
        ивенту. не работает для пользователей, которые уже состоят в этом ивенте.
      parameters:
      - description: Invite ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Присоединиться к ивенту.
      tags:
      - event
...
securityDefinitions:
  BearerAuth:
    description: Введите "Bearer {access_token}" (без кавычек)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"